feat: 예약이 존재하는 시간을 삭제하려할 때 예외 처리

+++ src/main/java/roomescape/service/ReservationTimeService.java
@@ -4,6 +4,7 @@ import java.util.List;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.web.server.ResponseStatusException;
+import roomescape.dao.ReservationDao;
 import roomescape.dao.ReservationTimeDao;
 import roomescape.dto.ReservationTimeRequest;
 import roomescape.dto.ReservationTimeResponse;
@@ -11,9 +12,11 @@ import roomescape.model.ReservationTime;
 
 @Service
 public class ReservationTimeService {
+    private final ReservationDao reservationDao;
     private final ReservationTimeDao reservationTimeDao;
 
-    public ReservationTimeService(ReservationTimeDao reservationTimeDao) {
+    public ReservationTimeService(ReservationDao reservationDao, ReservationTimeDao reservationTimeDao) {
+        this.reservationDao = reservationDao;
         this.reservationTimeDao = reservationTimeDao;
     }
 
@@ -24,6 +27,10 @@ public class ReservationTimeService {
     }
 
     public void deleteTime(Long id) {
+        int count = reservationDao.countByTimeId(id);
+        if(count != 0) {
+            throw new IllegalStateException("예약이 존재하여 삭제할 수 없습니다.");
+        }
         boolean isDeleted = reservationTimeDao.deleteById(id);
         if (!isDeleted) {
             throw new ResponseStatusException(HttpStatus.NOT_FOUND, "해당하는 id가 없습니다");
